+++ cloud-collect-demo/src/main/kotlin/com/windea/demo/cloudcollect/demo/validation/GroupA.kt
@@ -1,3 +1,0 @@
-package com.windea.demo.cloudcollect.demo.validation
-
-interface GroupA